refactor(api): reorganize tests outside src
