+++ apps/api/tests/integration/cattle/index.test.ts
@@ -1,16 +1,16 @@
 import type { AnyD1Database } from "drizzle-orm/d1";
 import { Hono } from "hono";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import cattleRoutes from "../../routes/cattle";
-import * as cattleService from "../../services/cattleService";
-import type { Bindings } from "../../types";
-import { mockCattle } from "../mocks/database";
+import cattleRoutes from "../../../src/routes/cattle";
+import * as cattleService from "../../../src/services/cattleService";
+import type { Bindings } from "../../../src/types";
+import { mockCattle } from "../../fixtures/database";
 
 // Mock the service module
-vi.mock("../../services/cattleService");
+vi.mock("../../../src/services/cattleService");
 
 // Mock JWT middleware
-vi.mock("../../middleware/jwt", () => ({
+vi.mock("../../../src/middleware/jwt", () => ({
 	jwtMiddleware: vi.fn((c, next) => {
 		c.set("jwtPayload", { userId: 1 });
 		return next();
