+++ apps/api/tests/unit/index.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import app from "../../index";
+import app from "../../src/index";
 
 describe("Main App", () => {
 	it("should export app instance", () => {
