+++ apps/api/tests/unit/lib/token.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { generateToken, hashPassword } from "../../lib/token";
+import { generateToken, hashPassword } from "../../../src/lib/token";
 
 // crypto.getRandomValuesをモック
 const mockGetRandomValues = vi.fn();
